+++ CHANGELOG.md
@@ -7,14 +7,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Added
 - Net energy content model: `NetEnergyModel::GrossEnergyFraction`
     - Parameter `forage.gross_energy`
+    - Parameter `hft.digestion.digestibility_multiplier`
     - Parameter `hft.digestion.k_fat`
     - Parameter `hft.digestion.k_maintenance`
     - Parameter `hft.digestion.me_coefficient`
 
 ### Removed
 - The old “default” net energy model `get_net_energy_content_default()`.
-    - Parameter `hft.digestion.efficiency` (less efficient digestion of hindgut
-      fermenters is now in `hft.digestion.me_coefficient`).
+- Parameter `hft.digestion.efficiency` (less efficient digestion of hindgut
+  fermenters is now in `hft.digestion.digestibility_multiplier`).
 
 ## 0.2.0 - 2019-12-06
 ### Added